s32 model ID to make the code cleaner

+++ include/types.h
@@ -54,7 +54,7 @@ typedef s16 Trajectory;
 typedef s16 PaintingData;
 typedef uintptr_t BehaviorScript;
 typedef u8 Texture;
-typedef u16 ModelID;
+typedef s32 ModelID;
 
 enum SpTaskState {
     SPTASK_STATE_NOT_STARTED,
